requirements.txt
Jinja2>=3.1
PyYAML>=6.0
markdown>=3.5
